Skip install of Qt 5.5 if Xcode not installed
While trying to run laptop on a new machine, I got the following error:

> A full installation of Xcode.app is required to compile this software.
> Installing just the Command Line Tools is not sufficient.

To truly fix this, we'd need to script the installation of Xcode, which
is beyond my skills at the moment. For now, I use a conditional in the
generated `Brewfile` to skip if Xcode is not installed.